Dev: teleport helpers
